scrapyrealestate/requirements.txt
###### Requirements without __version__ Specifiers ######
beautifulsoup4
PyYAML
art
scrapy
scrapy_useragents
scrapy-rotating-proxies
requests
pyTelegramBotAPI
requests_html
#pymongo
wheel
flask
unidecode
mysql-connector-python
#flask-sqlalchemy
#flask-marshmallow
#marshmallow-sqlalchemy
